moved column mapping and mock data to a new data mapping file
